+++ test/test_mediafile.py
@@ -291,7 +291,29 @@ class GenreListTestMixin:
         self.assertCountEqual(mediafile.genres, ["the genre", "another"])
 
 
-class ReadWriteTestBase(ArtTestMixin, GenreListTestMixin, _common.TempDirMixin):
+class LyricistListTestMixin:
+    """Tests access to the ``lyricists`` property as a list."""
+
+    def test_write_lyricist_list(self):
+        mediafile = self._mediafile_fixture("empty")
+        mediafile.lyricists = ["one", "two"]
+        mediafile.save()
+
+        mediafile = MediaFile(mediafile.filename)
+        self.assertCountEqual(mediafile.lyricists, ["one", "two"])
+
+    def test_write_lyricist_list_get_first(self):
+        mediafile = self._mediafile_fixture("empty")
+        mediafile.lyricists = ["one", "two"]
+        mediafile.save()
+
+        mediafile = MediaFile(mediafile.filename)
+        self.assertEqual(mediafile.lyricist, "one")
+
+
+class ReadWriteTestBase(
+    ArtTestMixin, GenreListTestMixin, LyricistListTestMixin, _common.TempDirMixin
+):
     """Test writing and reading tags. Subclasses must set ``extension``
     and ``audio_properties``.
 
@@ -868,6 +890,15 @@ class WMATest(ReadWriteTestBase, ExtendedImageStructureTestMixin, unittest.TestC
         mediafile = MediaFile(mediafile.filename)
         self.assertIn(mediafile.genre, ["one", "two"])
 
+    def test_write_lyricist_list_get_first(self):
+        # WMA does not preserve list order
+        mediafile = self._mediafile_fixture("empty")
+        mediafile.lyricists = ["one", "two"]
+        mediafile.save()
+
+        mediafile = MediaFile(mediafile.filename)
+        self.assertIn(mediafile.lyricist, ["one", "two"])
+
     def test_read_pure_tags(self):
         mediafile = self._mediafile_fixture("pure")
         self.assertEqual(mediafile.comments, "the comments")
@@ -1114,6 +1145,7 @@ class MediaFieldTest(unittest.TestCase):
                 "encoder",
                 "images",
                 "genres",
+                "lyricists",
                 "albumtype",
                 "artists",
                 "albumartists",