+++ test/test_mediafile.py
@@ -311,8 +311,32 @@ class LyricistListTestMixin:
         self.assertEqual(mediafile.lyricist, "one")
 
 
+class ComposerListTestMixin:
+    """Tests access to the ``composers`` property as a list."""
+
+    def test_write_composer_list(self):
+        mediafile = self._mediafile_fixture("empty")
+        mediafile.composers = ["one", "two"]
+        mediafile.save()
+
+        mediafile = MediaFile(mediafile.filename)
+        self.assertCountEqual(mediafile.composers, ["one", "two"])
+
+    def test_write_composer_list_get_first(self):
+        mediafile = self._mediafile_fixture("empty")
+        mediafile.composers = ["one", "two"]
+        mediafile.save()
+
+        mediafile = MediaFile(mediafile.filename)
+        self.assertEqual(mediafile.composer, "one")
+
+
 class ReadWriteTestBase(
-    ArtTestMixin, GenreListTestMixin, LyricistListTestMixin, _common.TempDirMixin
+    ArtTestMixin,
+    GenreListTestMixin,
+    LyricistListTestMixin,
+    ComposerListTestMixin,
+    _common.TempDirMixin,
 ):
     """Test writing and reading tags. Subclasses must set ``extension``
     and ``audio_properties``.
@@ -899,6 +923,15 @@ class WMATest(ReadWriteTestBase, ExtendedImageStructureTestMixin, unittest.TestC
         mediafile = MediaFile(mediafile.filename)
         self.assertIn(mediafile.lyricist, ["one", "two"])
 
+    def test_write_composer_list_get_first(self):
+        # WMA does not preserve list order
+        mediafile = self._mediafile_fixture("empty")
+        mediafile.composers = ["one", "two"]
+        mediafile.save()
+
+        mediafile = MediaFile(mediafile.filename)
+        self.assertIn(mediafile.composer, ["one", "two"])
+
     def test_read_pure_tags(self):
         mediafile = self._mediafile_fixture("pure")
         self.assertEqual(mediafile.comments, "the comments")
@@ -1146,6 +1179,7 @@ class MediaFieldTest(unittest.TestCase):
                 "images",
                 "genres",
                 "lyricists",
+                "composers",
                 "albumtype",
                 "artists",
                 "albumartists",